Fix incorrect FrameState stack type tracking for InterpOne
Summary:In FrameStateMgr::update(), if an InterpOne IR instruction has a type
parameter, we update the type info for the top stack slot of the
bytecode eval stack.  However, an InterpOne is given a type param for
its output even when it's not on the stack---e.g., if it's a local.

We need to check the output flags for the InterpOne'd HHBC instruction
before we try to clobber stack types.

This sometimes happens when the bytecode eval stack is empty, but it
was missed by our assertions because of an off-by-one error in stack
accounting---namely, we were 1-indexing the FrameState::stack vector.
When this bug came up when the BC eval stack was empty, we would
harmlessly write garbage to the 0th element of FrameState::stack, which
nobody else ever touched.  But when the BC eval stack was /not/ empty,
we would scribble over a previously correctly-typed stack slot.

Reviewed By: markw65

Differential Revision: D3036087

fb-gh-sync-id: 689ec0d443f6f1a3ad10cc0df881f13e08f8a12d
shipit-source-id: 689ec0d443f6f1a3ad10cc0df881f13e08f8a12d