Fix problem with errors containing no position
Summary:Fix for https://our.intern.facebook.com/intern/tasks/?q=1549737682016859&t=10355299 revealed a number of errors
in www, most of which can be suppressed by HH_FIXME, but three of which don't have attached file positions.

These can be traced to code in Typing_lenv, which passes Pos.none to the unifier.

Fix is to pass on the position from env.

Reviewed By: dlreeves

Differential Revision: D3063859

fb-gh-sync-id: ff54afcc816b60141712bfab06c6deb46c2c5044
shipit-source-id: ff54afcc816b60141712bfab06c6deb46c2c5044