Fix problem with lax checking of super constraints
Summary:There are a number of examples where "super" constraints on type parameters are not checked properly
by the Hack type checker, leading to a type loophole. This seems to be due to some special treatment
of super constraints in typing_generic_constraint.ml. I have removed this code - the example that
motivated it actually passes the type checker as expected.

Reviewed By: int3

Differential Revision: D3035399

fb-gh-sync-id: a5edbc4284d6af5357b0e6851da5ad2c2e41393f
shipit-source-id: a5edbc4284d6af5357b0e6851da5ad2c2e41393f