Deal with request-heap pointers in Transport class heirarchy.
Summary:Transport contains a Cell pointing to a callback function
in the PHP heap. Move it to ExecutionContext.

ReplayTransport contains an IniSettingMap: make it a root,
just to be safe; it only should have static contents.

XboxServer and PageletServer point to AsioExternalThreadEvent
instances which in contain pointers to WaitHandles in *other*
requests. added comments, but don't scan them. Threads only scan
pointers into their own heaps.

Reviewed By: jano

Differential Revision: D3054805

fb-gh-sync-id: 17345d23b008dfd7301b8b39c68fb99c371c2634
shipit-source-id: 17345d23b008dfd7301b8b39c68fb99c371c2634