Hack: make the HHI tarball generation umask-agnostic
Summary:This is a follow-up to D2834659, which aimed to make the HHI tarball
generation deterministic/reproducible. That commit unfortunately didn't
catch the case of a varying umask, which results in the tarball having
different modes for files/directories depending on the builder's umask.
Pass --mode to tar, to force the mode to a static one.
Closes https://github.com/facebook/hhvm/pull/6932

Reviewed By: Orvid

Differential Revision: D3088173

fb-gh-sync-id: 0c4e3b1163b8e4adb48ac3b9b2f2eb647ba0f77d
shipit-source-id: 0c4e3b1163b8e4adb48ac3b9b2f2eb647ba0f77d