Don't cache dynamic enum constants across requests
Summary:It's possible for a dynamic enum value to change between requests without the
enum itself changing. As we currently only clear the cache when the enum is
treadmilled this can lead to stale values persisting in the cache. Class* is
careful to use a request local cache for such values, so do the same in the
EnumCache.

Reviewed By: markw65

Differential Revision: D3088624

fb-gh-sync-id: 4e18db42df02259174bcf699f77f6dda1bfcb0f3
shipit-source-id: 4e18db42df02259174bcf699f77f6dda1bfcb0f3