Add type-index machinery to memory allocations
Summary:The type-scanners use a "type-index" to identify unique types stored in the
heap. Add the machinery to the memory allocation functions to record the
type-index of the type being allocated. req::malloc and family now take an
optional type-index parameter, which will be recorded in the heap. If the
type-index isn't provided, it receives a default "unknown-type" type-index,
which will trigger conservative scanning. Add additional *_noptr variants of the
allocation functions, used when the type isn't known, but we can assert the type
won't contain any pointers. These are ignored by the type-scanners. If one uses
the typed allocation functions (make_raw, etc), all of this is handled
automatically. So, going forward, using the typed allocations should be
preferred over the malloc functions.

Modify callers to req::malloc and family as necessary to use the *_noptr
variants, or the typed versions.

override-unit-failure

Reviewed By: edwinsmith

Differential Revision: D3026079

fb-gh-sync-id: 41333459f8cf06eb5dcc59da7f5808f0ae2972cc
shipit-source-id: 41333459f8cf06eb5dcc59da7f5808f0ae2972cc