Change AsyncGeneratorWaitHandle::m_generator to point to Object
Summary:Its clearest if heap->heap pointers are the usual data
types, for example ObjectData* instead of AsyncGenerator*
(the latter is a NativeData prefix type). This diff doesn't
change the raw-pointerness of the pointer.

Refactored c_AsyncGenerator::initialize() to do the work
in the constructor.

Remove unused file: ext_async-generator-inl.h

Reviewed By: jano

Differential Revision: D3082379

fb-gh-sync-id: 22bb42d5a269df876bb210aae967679700139ec0
fbshipit-source-id: 22bb42d5a269df876bb210aae967679700139ec0