Basic concurrent jitting
Summary:This diff adds an option to enable running most of the jit pipeline
concurrently in multiple threads. The only step that must be done while holding
the write lease is final code emission. Other than a whole bunch of code motion
and breaking up large functions, there are two major changes required to do
this:

First, LeaseHolder now has more states. Rather than "do I have the lease or
not?", users must ask "can I translate?" or "can I write to the translation
cache?" (the latter implies the former). Lease acquisition sites that support
doing something special in the canTranslate() case opt-in to this by passing a
Func* to LeaseHolder's constructor, which is used to ensure we only have one
translating thread per Func (for now. this may change in the future). This is a
disaster for code layout, as expected, and will be restricted to Profile
translations once I'm convinced we can correctly handle concurrent jitting in
all situations.

Second, getTranslation() now returns a TransResult rather than just a
TCA. TransResult contains either a TCA of the finished translation, or enough
state to finish the translation at a later time. This is also opt-in, so the
only way to get an unfinished translation is by setting the 'partial' member of
TransArgs to true. The one user of this for now is MCG::bindJccFirst(). It does
the first part of translation without requiring the write lease, then grabs the
write lease to handle its special alignment and fallthrough requirements before
final codegen.

PGO is not yet supported since ProfData needs some locking of its own rather
than relying on the write lease. That's next on my list.

Some miscellaneous related things:
- Rename TranslArgs to TransArgs for consistency with other structs.
- Remove TransArgs::setFuncBody and just do it manually at the two callsites
  that care.
- Use std::atomic<bool> instead of volatile in Lease.
- Remove unused gremlins from Lease.
- A blocking acquire of the write lease will never fail (unless there's a bug
  in our use of pthread_mutex_t) so remove checks for that.
- Improve server mode in test/run to start servers concurrently.

Reviewed By: mxw

Differential Revision: D2997746

fb-gh-sync-id: 00538fb14548939670d25fad5a88cc7ca97a11fc
fbshipit-source-id: 00538fb14548939670d25fad5a88cc7ca97a11fc