Fix archaeological curiosities in frame-state.cpp
Summary:There are some oddities in frame-state.cpp---places where we were
intentionally not using helpers, or doing checks, all of which seem to
relate to whether or not predicted types are allowed to go to Bottom.

I'm pretty sure they oughtn't be, so change the code to enforce it.
Doing so doesn't affect tests or perflab success, and also makes sense
reading the code, so it's probably correct.

Reviewed By: swtaarrs

Differential Revision: D3087548

fb-gh-sync-id: 65d5aceb05ce4b2cf709d6cbfb874e171bcb1acb
fbshipit-source-id: 65d5aceb05ce4b2cf709d6cbfb874e171bcb1acb