Compute list of all identifiers in auto-complete code lazily, speed up method/static method autocomplete 3x
Summary:Autocomplete path is probably our most (next to init) time-sensitive code path, yet it contains a diamond in the form of "iterate over all identifiers in codebase". It's used for global identifier autocomplete, and I guess it used to work fast enough 2+ years ago when Hack was new and codebase small.

I'm planning to optimize this thing more (see #10610556), but for now just pushing it out of the way of method/static method autocomplete where it is completely unnecessary, and looks like it was done only to give function a nice interface at the expense of what used to be a negligible perf impact.

Note that I already "optimized" it once, in D2835619, when changing sets to lists was also 3x speed up.

Reviewed By: ericlippert

Differential Revision: D3108966

fb-gh-sync-id: cf3bb6d53bf1cfb8aeb7f224cd4d1eeec69a0b09
fbshipit-source-id: cf3bb6d53bf1cfb8aeb7f224cd4d1eeec69a0b09